requirements.txt
django
psycopg2
djangorestframework
django-filter
requests
django-cors-headers
python-dateutil
